Out statement refactor
