+++ src/imp/types/builder.ts
@@ -1,15 +1,7 @@
-import { CompiledItem, OverpassOutputOptions, OverpassSettings, OverpassStatement } from "@/model";
+import { CompiledItem, OverpassSettings, OverpassStatement } from "@/model";
 
 export interface OverpassQueryBuilder {
 	buildSettings(settings: OverpassSettings): CompiledItem;
 
-	buildOptions(options: OverpassOutputOptions): CompiledItem;
-
-	buildStatements(statements: OverpassStatement[]): CompiledItem;
-
-	buildQuery(
-		settings: OverpassSettings,
-		options: OverpassOutputOptions,
-		statements: OverpassStatement[],
-	): CompiledItem;
+	buildQuery(settings: OverpassSettings, statements: OverpassStatement[]): CompiledItem;
 }
